Updating Table/LinHash/TupGen
Table-Added commented out options for different index types in
Table.java. Also added a cast to Comparable[] on line 445 to fix an
issue with using the Linear Hash as the index.
LinHash-Forced the key values to be positive. Cheap man's way to fix not
handling negative key values.
TupGen-The Teaching range select was taking up to 60 seconds per run so
I limited the number of runs for testing. The joins are also taking a
while, though not nearly 60 seconds.